refactor(gitlab-client): separate features by package due to quickly growing main package and model, making it difficult to read and understand

+++ ambassador-document-analyzer/src/main/kotlin/com/filipowm/ambassador/document/TextDetails.kt
@@ -2,7 +2,7 @@ package com.filipowm.ambassador.document
 
 data class TextDetails(
     val hash: String?,
-    val size: Int?,
+    val size: Long?,
     val path: String?,
     val content: String?
 )
